lib/lambda/requirements.txt
fastapi
uvicorn
mangum
c2pa-python